Created scripts that execute the code on the Slides
We talked about not having a great way to verify that our instances worked with the version of the ChefDK. We also have a hard time verifying if the code in the slides works. It is also sometimes useful to work through a problem on your own machine but you are too far behind to catch up.

These script files are tied to each section and will complete the exercises in each of their sections.

Running these will:

* Verify that the instance works with the code within the slide
* Allow an instructor or attendee to catch up

Introducing these scripts outside of the slide means that they are very likely to get out of date. I think they still provide value.