Updated Resources Section for Windows
We no longer have the learners install editors instead we immediately start teaching them about resources. The first activity is to create the hello.txt file through a recipe file. As a lab they create a recipe to remove that same file. Then to do something more impactful on Windows we walk them through disabling UAC with a registry_key resource. As a lab they complete the disabling of UAC.