Updated Cookbooks Section for Windows
The setup recipe is now the disable-uac recipe that is moved into the workstation cookbook. The introduction material was updated to be little less awkward with the "questions that they learner may have". The learner is still taught git. They are given a challenge to setup IIS - similar to the apache cookbook.