Updated the Testing Cookbooks for Windows
Testing for the windows world requires that a template yaml file be placed on the images that the learner will be copying over and using on that system. This will allow them to spin up new EC2 instances with Test Kitchen. The workstation machine itself needs to have the appropriate role defined on the system to allow the machine to spin up new instances without the presence of keys.

Testing is done through InSpec and not ServerSpec. At the moment with ChefDK 0.10.0 there are some workaround commands that have to be run to place the test suite in the proper state.