Updated Deatils About The System module for Windows
Updated to only work through defining node attributes for the iis-demo cookbook. I could not find another activity that did not seem too much of a repetition of too many steps. Namely I could not find an activity that mapped well to the message of the day (MOTD) in the workstation cookbook for linux. Instead I opted to leave no additional activity.