Updated the modules of day 2 (09-14)
All of the day 2 modules were updated to have more windows friendly filepaths.
All mentions of SSH were replaced with WINRM.
Updated any situation where it made sense to talk about Windows variation. This happens at one moment when we talk about Package, Template, and Service being the main resources of Config Management. This is not the case in Windows. It is more powershell_Script, registry_key, and service.