Removed extra version tracked in the Repository
Git contains all the history of all the content. Tracking all the versions
of the documents as additional files creates additional fies. If someone
wanted an older version they would want to move through the history of the
git repository to find the version they are after (probably through tags)
and then retrieve the state of the document then.