Removed version number from Guide Documents
The version number is not necessary. The version of the documents is known
by the repository version. Adding the version runs the risk of creating
confusion or having the repository track far too many versions. As these
documents are more released artifacts I am changing them here to not have
their version numbers and instead be representative of the current state
of the repository. We would then rely on git tags to find the version that
someone is after if it is not the latest.