Update all modules to consistent use of terminal slide FIX #171
The templates have become very different over time. Some of the templates do not even have the 'remote' command slide. For the ones that did I started to switch the layout of them all to 'local' with the little computer. However, I changed my strategry and simply edited the 'remote' one in the master to look like the 'local' one we want.

I also went through and removed all the 'chef hat' symbols on the very basic slides. It provided no value.